Fixes for payment widget work (#287)
* load and return data about workspace, plan and user to client

* fix param name

* fix error with circular deps

* add tariff plan id to response

* Long to Int

* log some data

* log some data

* add body-parser

* fix errors

* fix error code

* add docs

* fix integration tests

* fixes

* fix fetching operations

* add PENNY_MULTIPLIER

* review fixes

* fix integration tests

* fix invoice id

+++ src/typeDefs/plans.ts
@@ -18,7 +18,7 @@ export default gql`
     """
     Monthly charge for plan
     """
-    monthlyCharge: Long!
+    monthlyCharge: Int!
 
     """
     Events limit for plan
